mz/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
Django==4.2.13
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
install==1.3.5
pillow==10.3.0
PyJWT==2.8.0
python-bidi==0.4.2
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.0
tzdata==2024.1
